Major update to the LaTeX generation routine
Most importantly: I added a derived class for custom formatting differing from Pod::PseudoPod::LaTeX. This is actually a lot of stuff, since P::P::L does many things unflexibly which are better solved in the LaTeX preamble. Also added a TODO file for future changes, all of those (and of the ones in this commit) are subject to discussion.

Other changes in no particular order:

* Extracted LaTeX preamble to a separate file
* Enabled microtyping (http://en.wikipedia.org/wiki/Microtypography)
* Re-activated ligatures, why were they disabled in first place?
* Included line-numbering in code blocks
* Added proper page headings
* Changed margins
* Changed the layout of the authors list on the title page
* Probably a lot of small changes I've already forgotten about