[FIX] GeekCook example works on rakudo now
GeekCook.new hat two issues. First, the return self.bless got the wrong
parameters. Thus, it did not return the right object.

Also, it pushes onto %params<cookbooks> which might be undef. In rakudo this
does not enough autovivification (whatever the verb is) to make this push
possible. Therefore, the hash element needs to be initialised. It's not nice
to have a workaround for unimplemented functions but having an example that
does not work is certainly worse.